demo/requirements.txt
Django>=2.0.8
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
../.
